requirements.txt
# Python version requirement
# Ensure you are using Python 3.10.12 before installing dependencies

matplotlib==3.10.1
numpy==1.24.4
plotly==6.0.1
pandas==2.2.3
fpdf2==2.8.2
Pillow==9.0.1
